move DisposeTrackingBufferPool to its own file
